+++ lib/react_render/render_error.ex
@@ -1,0 +1,7 @@
+defmodule ReactRender.RenderError do
+  @moduledoc """
+  Error when unable to render given component
+  """
+
+  defexception message: nil, stack: nil
+end